Renamed features folders to tabs folders.
